+++ tests/test_linkrot.py
@@ -52,5 +52,19 @@ def test_single_link_is_dead(client) -> None:
 
 def test_single_link_is_web_archive(client) -> None:
     """Ensure a dead link is flagged as a web archive link."""
-    # TODO: Fill in this test
-    assert True is True
+    creation = client.post(
+        helpers.authed_request("/", auth=helpers.VALID_AUTH),
+        json=helpers.item_web_archive_url(),
+    )
+    item_id = helpers.from_json(creation.data)["id"]
+    for _ in range(3):
+        response = client.post(
+            helpers.authed_request("/", "linkrot", item_id, auth=helpers.VALID_AUTH),
+        )
+    response_data = helpers.from_json(response.get_data(as_text=True))
+    assert response.status_code == codes.OK
+    assert response_data["id"] == item_id
+    assert response_data["url"] == helpers.item_web_archive_url()["url"]
+    assert response_data["result"]["times_failed"] == 0
+    assert response_data["result"]["is_dead"] is False
+    assert response_data["result"]["is_web_archive"] is True